Add an documentation to the simple MergedEquippable
This commit adds a documentation to simple MergedEquippable as well as
modifies its deploy script to simply deploy it. User journey script has
been added to replace the showcase of simple MergedEquippable usage.

Custom script for running the simple MergedEquippable user journey has
been added.

Simple MergedEquippable example smart contracts have been moved into
their own folder, so that the documentation of each architecture can be
separate from the main README.
