+++ contracts/MergedEquippable/SimpleEquippable.sol
