+++ contracts/README.md
@@ -7,3 +7,4 @@ The examples covered are:
 1. [Nesting](./Nesting/README.md) examines the Nesting RMRK lego.
 2. [MultiResource](./MultiResource/README.md) examines the MultiResource RMRK lego.
 3. [NestingMultiResource](./NestingMultiResource/) examines the Nesting and MultiResource RMRK legos operating together.
+4. [MegredEquippable](./MergedEquippable/README.md) examines the Merged equippable RMRK lego formation.
